Show initial array of objects from redux store
