+++ src/index.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
-//import ReactDOM from 'react-dom';
 import * as ReactDOMClient from 'react-dom/client';
 import './index.css';
 import App from './App';
+import { store } from './store/store';
+import { Provider } from 'react-redux';
 
 const container = document.getElementById('root') as Element
 const root = ReactDOMClient.createRoot(container)
 
 root.render(
   <React.StrictMode>
-    <App />
+    <Provider store={store}>
+      <App />
+    </Provider>
   </React.StrictMode>
 );
