+++ src/store/store.ts
@@ -1,15 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
-
-import knowledgeItemReducer from './knowledgeItem';
+import knowledgeItemReducer from './knowledgeItemSlice';
 
 export const store = configureStore({
   reducer: {
-    knowledgeItem: knowledgeItemReducer
+    knowledgeItem: knowledgeItemReducer,
   }
 })
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
 export type RootState = ReturnType<typeof store.getState>
-
 // Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
 export type AppDispatch = typeof store.dispatch
